Free-node UI (node ref-count tracking) actually shows refcounts now.
Discovered a couple of rules that I need to implement:

1. If a term w/Loc is put in a neg node and the refcount of that node's loc is
   zero, decrement the refcount of the term's loc.

2. If a term w/Loc is put in a pos node and the corresponding neg node's term
   is NUL/TAKEN, and the refcount of the pos node's loc is zero, decrement the
    refcount of the term's loc.

I believe it is possible that with only these two rules, I can syncrhonize the
free (orange) state of nodes displayed in "ref rects", with the free-state of
node addresses displayed in the free-node (refcount) UI.

Ref rects are cheating - they have knowledge of all memory operations in the
logfile and thus know exactly at what point no subsequent operations will occur.

Free-node UI has no such knowledge, it is simply keeping track of the 'loc'
field of the terms of every memory operation as it is encountered, and will use
a couple other heuristics as detailed above.

If this works, it means ~100% node recycling for a relatively small performance
hit on single threaded. Multi-threaded may be a little trickier, but potentially
faster (and a lot simpler) than any method maintaining a graph of disconnected
nodes.